pyproject.toml
[project]
name = "ics-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.3.0",
    "icalendar>=7.0.2",
    "pydantic>=2.12.5",
    "pypdf>=6.7.4",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
